requirements.txt
boto3
pydub
pillow